modules/module3/old_content/code/scraping_data/requirements.txt
lxml==3.6.0
MySQL-python==1.2.5
pytz==2016.3
requests==2.9.1
retrying==1.3.3
six==1.10.0
SQLAlchemy==1.0.12
tqdm==4.5.0
tzlocal==1.2.2
psycopg2==2.6.1
python-dateutil==2.5.3
beautifulsoup4==4.4.1

